backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
pytest==6.2.4
pytest-django==4.4.0
gunicorn==20.1.0
python-dotenv==1.0.0
django-filter==22.1
drf-extra-fields==3.4.0